[DIC] Move PrependExtensionInterface to the Extension namespace
